No Item View added To the list View

+++ ToDoList/Views/ListView.swift
@@ -5,26 +5,31 @@ struct ListView: View {
     @EnvironmentObject var listViewModel: ListViewModel
     
     var body: some View {
-        ZStack {
-            List {
-                ForEach(listViewModel.items) { item in
-                    ListRowView(item: item)
-                        .onTapGesture {
-                            withAnimation(.linear) {
-                                // Handle tap gesture
-                                listViewModel.updateItem(item: item)
-                            }
+        ZStack{
+            if listViewModel.items.isEmpty{
+                NoItemsView()
+            }
+            else{
+                    List {
+                        ForEach(listViewModel.items) { item in
+                            ListRowView(item: item)
+                                .onTapGesture {
+                                    withAnimation(.linear) {
+                                        // Handle tap gesture
+                                        listViewModel.updateItem(item: item)
+                                    }
+                                }
                         }
-                }
-                .onDelete(perform: listViewModel.deleteItem) // Swipe to delete
-                .onMove(perform: listViewModel.moveItem) // Reorder items
-            }.listStyle(.plain)
-            .navigationTitle("ToDo List 📝") // Set navigation title
+                        .onDelete(perform: listViewModel.deleteItem) // Swipe to delete
+                        .onMove(perform: listViewModel.moveItem) // Reorder items
+                    }.listStyle(.plain)
+            }
+            
+        }.navigationTitle("ToDo List 📝") // Set navigation title
             .navigationBarItems(
                 leading: EditButton(), // Edit button for reordering and deletion
                 trailing: NavigationLink("Add", destination: AddView()) // Navigation link to AddView for adding new items
             )
-        }
     }
 }
